--- a/.gitignore
+++ b/.gitignore
@@ -19,6 +19,4 @@ docs/
 !.vscode/extensions.json
 !.vscode/*.code-snippets
 
-.github/workflows 
-
 *lcov.info