--- a/foundry.toml
+++ b/foundry.toml
@@ -2,6 +2,8 @@
 src = 'src'
 out = 'out'
 libs = ['lib']
-
+optimizer = true
+optimizer_runs = 100
+via_ir = false
 
 # See more config options https://github.com/foundry-rs/foundry/tree/master/config